fix delete User and TaskStatus if tasks not empty

+++ src/main/java/hexlet/code/controller/TaskStatusController.java
@@ -51,7 +51,11 @@ public class TaskStatusController {
     @DeleteMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
-        taskStatusRepository.deleteById(id);
+
+        if (taskStatusRepository.findById(id).get().getTasks().isEmpty()) {
+            taskStatusRepository.deleteById(id);
+        }
+
     }
 
     @PostMapping(path = "")
